backend/requirements.txt
Django==5.2.6
# REST framework
djangorestframework==3.16.1
# Simple JWT auth
djangorestframework-simplejwt==5.5.1
# CORS
django-cors-headers==4.6.0
# Env loader
python-dotenv==1.0.1
# MySQL driver
mysqlclient==2.2.7
